+++ CHANGELOG.md
@@ -22,7 +22,7 @@ have been replaced with `VideoLiveInfo`
 - Added `Serializable`
 - Changed `Video` to extend `Serializable`
 - All models are now const
-- Add String field `code` to enum `Language`, `Includes`, `Order`, `ChannelSort`, `VideoStatus`
+- Add String field `code` to enum `Language`, `Includes`, `Order`, `ChannelSort`, `VideoStatus`, `VideoSort`
 
 ## 0.3.1